Relocate reusable WzFrontendImageButton
