+++ src/titleui/campaign.cpp
@@ -45,6 +45,7 @@
 #include "../modding.h"
 #include "../version.h"
 #include "widgets/infobutton.h"
+#include "widgets/frontendimagebutton.h"
 #include "widgets/advcheckbox.h"
 
 typedef WzAdvCheckbox WzCampaignTweakOptionToggle;
@@ -793,178 +794,6 @@ public:
 	bool currentValue = false;
 	bool editable = true;
 };
-
-class WzFrontendImageButton : public W_BUTTON
-{
-protected:
-	WzFrontendImageButton() {}
-public:
-	static std::shared_ptr<WzFrontendImageButton> make(optional<UWORD> frontendImgID)
-	{
-		class make_shared_enabler: public WzFrontendImageButton {};
-		auto widget = std::make_shared<make_shared_enabler>();
-		widget->setImage(frontendImgID);
-		return widget;
-	}
-	void setImage(optional<UWORD> frontendImgID);
-	void setImageDimensions(int imageSize);
-	void setPadding(int horizontalPadding, int verticalPadding);
-public:
-	void setString(WzString string) override;
-	int32_t idealWidth() override;
-	int32_t idealHeight() override;
-protected:
-	void display(int xOffset, int yOffset) override;
-private:
-	void recalcIdealWidth();
-private:
-	WzText wzText;
-	optional<UWORD> frontendImgID = nullopt;
-	int imageDimensions = 16;
-	int cachedIdealWidth = 0;
-	int lastWidgetWidth = 0;
-	int horizontalPadding = 5;
-	int verticalPadding = 3;
-};
-
-void WzFrontendImageButton::setImage(optional<UWORD> newFrontendImgID)
-{
-	frontendImgID = newFrontendImgID;
-	recalcIdealWidth();
-}
-
-void WzFrontendImageButton::setImageDimensions(int imageSize)
-{
-	if (imageSize == imageDimensions)
-	{
-		return;
-	}
-	imageDimensions = imageSize;
-	recalcIdealWidth();
-}
-
-void WzFrontendImageButton::setPadding(int newHorizPadding, int newVertPadding)
-{
-	horizontalPadding = newHorizPadding;
-	verticalPadding = newVertPadding;
-	recalcIdealWidth();
-}
-
-void WzFrontendImageButton::setString(WzString string)
-{
-	W_BUTTON::setString(string);
-	recalcIdealWidth();
-}
-
-void WzFrontendImageButton::recalcIdealWidth()
-{
-	cachedIdealWidth = horizontalPadding;
-	if (frontendImgID.has_value() && imageDimensions > 0)
-	{
-		cachedIdealWidth += imageDimensions + horizontalPadding;
-	}
-	auto textWidth = iV_GetTextWidth(pText, FontID);
-	if (textWidth > 0)
-	{
-		cachedIdealWidth += textWidth + horizontalPadding;
-	}
-}
-
-int32_t WzFrontendImageButton::idealWidth()
-{
-	return cachedIdealWidth;
-}
-
-int32_t WzFrontendImageButton::idealHeight()
-{
-	return (verticalPadding * 2) + std::max<int32_t>(imageDimensions, iV_GetTextLineSize(FontID));
-}
-
-void WzFrontendImageButton::display(int xOffset, int yOffset)
-{
-	int x0 = x() + xOffset;
-	int y0 = y() + yOffset;
-	int w = width();
-
-	bool isDown = (getState() & (WBUT_DOWN | WBUT_LOCK | WBUT_CLICKLOCK)) != 0;
-	bool isDisabled = (getState() & WBUT_DISABLE) != 0;
-	bool isHighlight = !isDisabled && ((getState() & WBUT_HIGHLIGHT) != 0);
-	bool hasImage = frontendImgID.has_value() && imageDimensions > 0;
-
-	PIELIGHT borderColor = (isDown) ? WZCOL_TEXT_DARK : ((isHighlight) ? WZCOL_TEXT_BRIGHT : WZCOL_TEXT_MEDIUM);
-	PIELIGHT textColor = (isHighlight) ? WZCOL_TEXT_BRIGHT : WZCOL_TEXT_MEDIUM;
-	if (isDisabled)
-	{
-		borderColor.byte.a = (borderColor.byte.a / 2);
-		textColor.byte.a = (textColor.byte.a / 2);
-	}
-
-	// Display outer border
-	int boxX0 = x0;
-	int boxY0 = y0;
-	int boxX1 = boxX0 + width() + 1;
-	int boxY1 = boxY0 + height() + 1;
-	iV_Box(boxX0, boxY0, boxX1, boxY1, borderColor);
-
-	if (!pText.isEmpty())
-	{
-		// Display text
-		iV_fonts fontID = wzText.getFontID();
-		if (fontID == font_count || lastWidgetWidth != width() || wzText.getText() != pText)
-		{
-			fontID = FontID;
-		}
-		wzText.setText(pText, fontID);
-
-		int availableButtonTextWidth = w - (horizontalPadding * 2) - ((hasImage) ? (imageDimensions + horizontalPadding) : 0);
-		if (wzText.width() > availableButtonTextWidth)
-		{
-			// text would exceed the bounds of the button area
-			// try to shrink font so it fits
-			do {
-				if (fontID == font_small || fontID == font_bar)
-				{
-					break;
-				}
-				auto result = iV_ShrinkFont(fontID);
-				if (!result.has_value())
-				{
-					break;
-				}
-				fontID = result.value();
-				wzText.setText(pText, fontID);
-			} while (wzText.width() > availableButtonTextWidth);
-		}
-		lastWidgetWidth = width();
-
-		// Draw the main text
-		int textX0 = x0 + horizontalPadding + ((hasImage) ? imageDimensions + horizontalPadding : 0);
-		int textY0 = static_cast<int>(y0 + (height() - wzText.lineSize()) / 2 - float(wzText.aboveBase()));
-
-		const int maxTextDisplayableWidth = w - (horizontalPadding * 2);
-		int maxDisplayableMainTextWidth = maxTextDisplayableWidth;
-		bool isTruncated = maxDisplayableMainTextWidth < wzText.width();
-		if (isTruncated)
-		{
-			maxDisplayableMainTextWidth -= (iV_GetEllipsisWidth(wzText.getFontID()) + 2);
-		}
-		wzText.render(textX0, textY0, textColor, 0.0f, maxDisplayableMainTextWidth);
-		if (isTruncated)
-		{
-			// Render ellipsis
-			iV_DrawEllipsis(wzText.getFontID(), Vector2f(textX0 + maxDisplayableMainTextWidth + 2, textY0), textColor);
-		}
-	}
-
-	if (frontendImgID.has_value())
-	{
-		// Display the image to the left, vertically-centered
-		int imgPosX0 = x0 + horizontalPadding;
-		int imgPosY0 = y0 + (height() - imageDimensions) / 2;
-		PIELIGHT imgColor = textColor;
-		iV_DrawImageFileAnisotropicTint(FrontImages, frontendImgID.value(), imgPosX0, imgPosY0, Vector2f(imageDimensions, imageDimensions), imgColor);
-	}
-}
 
 class WzCampaignTweakOptionsEditForm : public WIDGET
 {
